requirements.txt
pandas>=2.2.0
numpy>=1.26.0
faker>=21.0.0
python-dateutil>=2.8.2
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
pydantic>=2.5.3
python-multipart>=0.0.6
email-validator>=2.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.3
pytest-asyncio>=0.23.2
httpx>=0.26.0
redis>=5.0.1
